+++ examples/mindshare/build.rs
@@ -28,6 +28,8 @@ fn main() {
         .crate_provides("capnp", [0xa93fc509624c72d9])
         .file(capnp_dir.join("system.capnp"))
         .file(capnp_dir.join("routing.capnp"))
+        .file(capnp_dir.join("auth.capnp"))
+        .file(capnp_dir.join("membrane.capnp"))
         .file(capnp_dir.join("stem.capnp"))
         .file(capnp_dir.join("http.capnp"))
         .run()
@@ -55,7 +57,7 @@ fn main() {
         .expect("write schema bytes");
 
     // Cargo rebuild triggers
-    for schema in &["system", "routing", "stem", "http"] {
+    for schema in &["system", "routing", "auth", "membrane", "stem", "http"] {
         println!(
             "cargo:rerun-if-changed={}",
             capnp_dir.join(format!("{schema}.capnp")).display()